+++ R/zzz.R
@@ -1,0 +1,24 @@
+# eulerr: Area-Proportional Euler and Venn Diagrams with Circles or Ellipses
+# Copyright (C) 2018 Johan Larsson <johanlarsson@outlook.com>
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+.eulerr_env <- new.env(parent = emptyenv())
+assign("options", list(), envir = .eulerr_env)
+
+.onLoad <- function(libname, pkgname)
+  eulerr_options(eulerr_default_options())
+
+.onUnload <- function(libpath)
+  library.dynam.unload("eulerr", libpath)